src/Backend/requirements.txt
Flask==3.0.0
pymongo[srv]==4.6.1
Flask-SQLAlchemy==3.1.1
Flask_JWT_Extended==4.6.0
python_decouple==3.8
flask_restx==1.3.0
flask-mongoengine==1.0.0
dnspython==2.0.0
email-validator==2.1.0
flask-pymongo==2.3.0
python-dotenv==1.0.0
python_decouple==3.8
flask_cors==4.0.0
flask-talisman==1.1.0
cffi==1.16.0
cryptography==42.0.1
pycparser==2.21
pyopenssl==24.0.0
bcrypt==4.1.2
six==1.16.0
flask_session==0.6.0